Improve playground Lighthouse loading
