+++ src/components/playground/Playground.tsx
@@ -1,12 +1,7 @@
-import React, {useEffect, useMemo, useState} from "react"
-import {GraphiQL} from "graphiql"
-import {analyticsHandler, isValidURL, sendConversionEvent} from "@site/src/utils"
-import {CookiePreferenceCategory, playgroundAdsConversionId} from "@site/src/constants"
-import "graphiql/graphiql.css"
-import "../../css/graphiql.css"
-import {type FetcherParams, FetcherOpts} from "@graphiql/toolkit"
-import {useCookieConsent} from "@site/src/utils/hooks/useCookieConsent"
-import {createGraphiQLFetcher} from "@graphiql/create-fetcher"
+import React, {Suspense, lazy, useEffect, useState} from "react"
+import {isValidURL} from "@site/src/utils"
+
+const GraphiQLPanel = lazy(() => import("./GraphiQLPanel"))
 
 const useDebouncedValue = (inputValue: string, delay: number) => {
   const [debouncedValue, setDebouncedValue] = useState(inputValue)
@@ -33,9 +28,6 @@ const Playground = () => {
   )
   const [inputValue, setInputValue] = useState<string>(initialApiEndpoint.toString())
 
-  const {getCookieConsent} = useCookieConsent()
-  const cookieConsent = getCookieConsent()
-
   const debouncedApiEndpoint = useDebouncedValue(inputValue, 500)
   const apiEndpointInputClasses = `border border-solid border-tailCall-border-light-500 rounded-lg font-space-grotesk h-11 w-[100%]
     p-SPACE_04 text-content-small outline-none focus:border-x-tailCall-light-700`
@@ -46,37 +38,7 @@ const Playground = () => {
     }
   }, [debouncedApiEndpoint])
 
-  const graphQLFetcher = async (graphQLParams: FetcherParams, opts?: FetcherOpts) => {
-    if (apiEndpoint.toString().trim() === "") {
-      return Promise.resolve({})
-    }
-    analyticsHandler("GraphQL", "tc_fetch_query", apiEndpoint.toString())
-    sendConversionEvent(playgroundAdsConversionId)
-
-    const fetcher = createGraphiQLFetcher({url: apiEndpoint.toString()})
-    return fetcher(graphQLParams, opts)
-  }
-
-  const emptyGraphiqlStorageObject = {
-    getItem: (): null => null,
-    setItem: (): void => undefined,
-    removeItem: (): void => undefined,
-    clear: (): void => undefined,
-    length: 0,
-  }
-
-  const graphiqlStorage = useMemo(() => {
-    if (
-      cookieConsent?.accepted &&
-      (!cookieConsent?.preferences || cookieConsent?.preferences?.includes(CookiePreferenceCategory.PREFERENCE))
-    ) {
-      // Defaults to local storage
-      return undefined
-    }
-
-    // Block storing graphiql data in local storage if user denies cookie consent
-    return emptyGraphiqlStorageObject
-  }, [cookieConsent])
+  const endpointUrl = apiEndpoint.toString().trim()
 
   return (
     <div className="min-h-[90vh]">
@@ -93,11 +55,11 @@ const Playground = () => {
             />
           </div>
           <div className="flex my-SPACE_03">
-            <GraphiQL fetcher={graphQLFetcher} storage={graphiqlStorage}>
-              <GraphiQL.Logo>
-                <></>
-              </GraphiQL.Logo>
-            </GraphiQL>
+            {endpointUrl !== "" && (
+              <Suspense fallback={null}>
+                <GraphiQLPanel apiEndpoint={endpointUrl} />
+              </Suspense>
+            )}
           </div>
         </div>
       )}
