+++ src/components/playground/index.tsx
@@ -1,6 +1,5 @@
 import React from "react"
 import Playground from "./Playground"
-import Announcement from "@site/src/components/shared/Announcement"
 
 const PlaygroundPage = (): JSX.Element => {
   return (
