+++ src/pages/playground/app.tsx
@@ -1,15 +1,15 @@
 import React, {useEffect} from "react"
 import ReactGA from "react-ga4"
 import Layout from "@theme/Layout"
-import PlaygroundPage from "../components/playground"
+import PlaygroundPage from "../../components/playground"
 import {useLocation} from "@docusaurus/router"
-import {PageDescription, PageTitle} from "../constants/titles"
+import {PageDescription, PageTitle} from "../../constants/titles"
 
-const Playground = () => {
+const PlaygroundApp = () => {
   const location = useLocation()
 
   useEffect(() => {
-    ReactGA.send({hitType: "pageview", page: location.pathname, title: "Playground Page"})
+    ReactGA.send({hitType: "pageview", page: location.pathname, title: "Playground App Page"})
   }, [])
 
   return (
@@ -19,4 +19,4 @@ const Playground = () => {
   )
 }
 
-export default Playground
+export default PlaygroundApp
