+++ static/playground/index.html
@@ -1,0 +1,204 @@
+<!doctype html>
+<html lang="en">
+  <head>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta
+      name="description"
+      content="Open Tailcall's GraphQL playground with a server endpoint."
+    />
+    <title>Tailcall Playground</title>
+    <link rel="icon" href="/images/favicon.ico" />
+    <style>
+      :root {
+        color-scheme: dark;
+        --bg: #080a0f;
+        --panel: #11151d;
+        --line: #283241;
+        --text: #f5f7fb;
+        --muted: #aeb8c8;
+        --accent: #f9c846;
+        --accent-ink: #14120a;
+      }
+
+      * {
+        box-sizing: border-box;
+      }
+
+      body {
+        margin: 0;
+        min-height: 100vh;
+        background: var(--bg);
+        color: var(--text);
+        font-family:
+          Inter, ui-sans-serif, system-ui, -apple-system, BlinkMacSystemFont, "Segoe UI",
+          sans-serif;
+      }
+
+      main {
+        display: grid;
+        min-height: 100vh;
+        place-items: center;
+        padding: 32px 16px;
+      }
+
+      section {
+        width: min(720px, 100%);
+      }
+
+      .brand {
+        display: inline-flex;
+        align-items: center;
+        gap: 10px;
+        color: var(--accent);
+        font-size: 14px;
+        font-weight: 700;
+        letter-spacing: 0;
+        margin-bottom: 22px;
+      }
+
+      .mark {
+        display: grid;
+        width: 28px;
+        height: 28px;
+        place-items: center;
+        border: 1px solid var(--accent);
+        border-radius: 6px;
+      }
+
+      h1 {
+        font-size: clamp(38px, 7vw, 76px);
+        line-height: 0.96;
+        letter-spacing: 0;
+        margin: 0 0 18px;
+      }
+
+      p {
+        color: var(--muted);
+        font-size: 18px;
+        line-height: 1.6;
+        margin: 0 0 28px;
+        max-width: 58ch;
+      }
+
+      form {
+        display: grid;
+        grid-template-columns: 1fr auto;
+        gap: 10px;
+        padding: 10px;
+        background: var(--panel);
+        border: 1px solid var(--line);
+        border-radius: 8px;
+      }
+
+      label {
+        position: absolute;
+        width: 1px;
+        height: 1px;
+        overflow: hidden;
+        clip: rect(0 0 0 0);
+        white-space: nowrap;
+      }
+
+      input,
+      button,
+      a {
+        font: inherit;
+      }
+
+      input {
+        width: 100%;
+        min-width: 0;
+        height: 48px;
+        color: var(--text);
+        background: #090c12;
+        border: 1px solid transparent;
+        border-radius: 6px;
+        outline: none;
+        padding: 0 14px;
+      }
+
+      input:focus {
+        border-color: var(--accent);
+      }
+
+      button,
+      .skip {
+        display: inline-flex;
+        align-items: center;
+        justify-content: center;
+        min-height: 48px;
+        border-radius: 6px;
+        border: 0;
+        padding: 0 18px;
+        font-weight: 700;
+        text-decoration: none;
+      }
+
+      button {
+        background: var(--accent);
+        color: var(--accent-ink);
+        cursor: pointer;
+      }
+
+      .skip {
+        width: max-content;
+        margin-top: 14px;
+        color: var(--text);
+        background: transparent;
+        border: 1px solid var(--line);
+      }
+
+      @media (max-width: 640px) {
+        main {
+          place-items: start;
+          padding-top: 56px;
+        }
+
+        form {
+          grid-template-columns: 1fr;
+        }
+
+        button,
+        .skip {
+          width: 100%;
+        }
+      }
+    </style>
+  </head>
+  <body>
+    <main>
+      <section>
+        <div class="brand" aria-label="Tailcall">
+          <span class="mark">&gt;</span>
+          Tailcall
+        </div>
+        <h1>GraphQL Playground</h1>
+        <p>
+          Connect to a GraphQL endpoint without loading the full editor until you are ready.
+        </p>
+        <form id="playground-form" action="/playground/app/" method="get">
+          <label for="endpoint">GraphQL endpoint</label>
+          <input
+            id="endpoint"
+            name="u"
+            type="url"
+            inputmode="url"
+            autocomplete="url"
+            placeholder="https://api.example.com/graphql"
+          />
+          <button type="submit">Open</button>
+        </form>
+        <a class="skip" href="/playground/app/">Open empty playground</a>
+      </section>
+    </main>
+    <script>
+      const params = new URLSearchParams(window.location.search)
+      const endpoint = params.get("u")
+
+      if (endpoint) {
+        window.location.replace(`/playground/app/?u=${encodeURIComponent(endpoint)}`)
+      }
+    </script>
+  </body>
+</html>
